basic product page refacto
