+++ _dev/js/theme/core/product/index.js
@@ -1,6 +1,5 @@
 import updateProduct from '@js/theme/core/product/updateProduct';
-import DOMReady from '@js/theme/utils/DOMReady';
 
-DOMReady(() => {
+$(() => {
   updateProduct();
 });
