+++ _dev/js/theme/corejs/index.js
@@ -8,7 +8,6 @@ import './selectors';
 import './checkout';
 import './facets';
 import './listing';
-import './product';
 import './address';
 
 import { psShowHide } from './common';
